fix: Fixed bug that caused wrong lesson type selection
In the get_lessons() function under utils, it now uses the
models factory function of get_type_display() to print the
type of the lesson. Previously, it used a tuple, indexed into
by the returned int from the db, a pretty manual and erroneous
method, since the tuple is zero indexed. The lesson types
however, were either 1 or 2, both ints.